:recycle: refactor: lint 및 미사용 모듈 제거 (#28)
