+++ src/domain/family-member/dto/request/create-family-member.dto.ts
@@ -1,5 +1,4 @@
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { Long } from 'typeorm';
 
 export class CreateFamilyMemberDto {
   @IsNotEmpty()
