+++ src/module/family-schedule.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
-import { FamilyScheduleController } from '../domain/family-schedule/family-schedule.controller';
-import { FamilyScheduleService } from '../domain/family-schedule/family-schedule.service';
+import {
+  FamilyScheduleController,
+  FamilyScheduleService,
+} from '../domain/family-schedule';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import {Family, FamilySchedule} from '../infra/entities';
+import { Family, FamilySchedule } from '../infra/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FamilySchedule, Family])],
