+++ src/module/photo.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PhotoController } from '../domain/photo/photo.controller';
-import { PhotoService } from '../domain/photo/photo.service';
+import { PhotoController, PhotoService } from '../domain/photo';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Family, Photo } from '../infra/entities';
