+++ src/module/post.module.ts
@@ -1,8 +1,7 @@
 import { Module } from '@nestjs/common';
-import { PostController } from '../domain/post/post.controller';
-import { PostService } from '../domain/post/post.service';
+import { PostController, PostService } from '../domain/post';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import {FamilyMember, Post} from '../infra/entities';
+import { FamilyMember, Post } from '../infra/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FamilyMember, Post])],
