+++ src/module/user.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserService } from '../domain/user';
-import { UserController } from '../domain/user';
+import { UserController, UserService } from '../domain/user';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../infra/entities';
 import { AuthModule } from '../auth';
